Implement SVG outlines and sideloading support with 100% coverage
- Added 'svg' outline type supporting inline path data (single or array).
- Implemented multi-path union and simplification for SVG outlines.
- Added sideloading of SVG files from 'outlines/' folder.
- Fixed horizontal mirroring and vertical Y-axis alignment for SVG paths.
- Added comprehensive unit and integration tests to achieve 100% code coverage.
- Ensured all 119 tests pass successfully.
