+++ src/outlines.js
@@ -1,16 +1,13 @@
 const m = require('makerjs')
-
 const u = require('./utils')
 const a = require('./assert')
 const o = require('./operation')
+const Point = require('./point')
 const prep = require('./prepare')
 const anchor = require('./anchor').parse
 const filter = require('./filter').parse
-
-const Point = require('./point')
+const injected_outlines = require('./outlines/index')
 const hulljs = require('hull')
-
-const outline_types = require('./outlines/index.js')
 
 const binding = (base, bbox, point, units) => {
 
@@ -361,13 +358,7 @@ const whats = {
     outline,
     path,
     svg,
-    hull,
-    ...outline_types
-}
-
-exports.inject_outline = (name, outline) => {
-    outline_types[name] = outline
-    whats[name] = outline
+    hull
 }
 
 const expand_shorthand = (config, name, units) => {
@@ -418,7 +409,7 @@ exports.parse = (config, points, units) => {
 
             // process keys that are common to all part declarations
             const operation = u[a.in(part.operation || 'add', `${name}.operation`, ['add', 'subtract', 'intersect', 'stack'])]
-            const what = a.in(part.what || 'outline', `${name}.what`, Object.keys(whats))
+            const what = a.in(part.what || 'outline', `${name}.what`, ['rectangle', 'circle', 'polygon', 'outline', 'path', 'hull', 'svg', ...Object.keys(injected_outlines)])
             const bound = !!part.bound
             const asym = a.asym(part.asym || 'source', `${name}.asym`)
 
@@ -447,7 +438,7 @@ exports.parse = (config, points, units) => {
             delete part.scale
 
             // a prototype "shape" maker (and its units) are computed
-            const [shape_maker, shape_units] = whats[what](part, name, points, outlines, units)
+            const [shape_maker, shape_units] = (whats[what] || injected_outlines[what])(part, name, points, outlines, units)
             const adjust = start => anchor(original_adjust || {}, `${name}.adjust`, points, start)(shape_units)
 
             // and then the shape is repeated for all where positions
@@ -485,4 +476,8 @@ exports.parse = (config, points, units) => {
     }
 
     return outlines
-}   
+}
+
+exports.inject_outline = (name, outline) => {
+    injected_outlines[name] = outline
+}
