+++ src/outlines/index.js
@@ -1,3 +1,2 @@
 module.exports = {
-  mx_hotswap: require('./mx_hotswap'),
 }
