--- test/cli/bundle/reference/outlines/hotswap.dxf
@@ -1,278 +1,0 @@
-0
-SECTION
-2
-HEADER
-9
-$INSUNITS
-70
-4
-0
-ENDSEC
-0
-SECTION
-2
-TABLES
-0
-TABLE
-2
-LTYPE
-0
-LTYPE
-72
-65
-70
-0
-2
-CONTINUOUS
-3
-______
-73
-0
-40
-0
-0
-LTYPE
-72
-65
-70
-0
-2
-DASHED
-3
-_ _ _ 
-73
-2
-40
-7.5
-49
-5
-49
--2.5
-0
-LTYPE
-72
-65
-70
-0
-2
-DOTTED
-3
-. . . 
-73
-2
-40
-1.5
-49
-0.5
-49
--1
-0
-ENDTAB
-0
-TABLE
-2
-LAYER
-0
-ENDTAB
-0
-ENDSEC
-0
-SECTION
-2
-ENTITIES
-0
-ARC
-8
-0
-6
-CONTINUOUS
-10
-8.4
-20
-6.58
-40
-2.05
-50
-180
-51
-540
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-12.95
-20
-2.8
-11
-15.55
-21
-2.8
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-15.55
-20
-2.8
-11
-15.55
-21
-0.2
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-15.55
-20
-0.2
-11
-12.95
-21
-0.2
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-12.95
-20
-0.2
-11
-12.95
-21
-2.8
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-0
-20
-5.34
-11
-2.6
-21
-5.34
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-2.6
-20
-5.34
-11
-2.6
-21
-2.74
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-2.6
-20
-2.74
-11
-0
-21
-2.74
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-0
-20
-2.74
-11
-0
-21
-5.34
-0
-ARC
-8
-0
-6
-CONTINUOUS
-10
-4.59
-20
-4.04
-40
-1.5
-50
-180
-51
-540
-0
-ARC
-8
-0
-6
-CONTINUOUS
-10
-10.94
-20
-1.5
-40
-1.5
-50
-180
-51
-540
-0
-ARC
-8
-0
-6
-CONTINUOUS
-10
-3.32
-20
-6.58
-40
-0.95
-50
-180
-51
-540
-0
-ARC
-8
-0
-6
-CONTINUOUS
-10
-13.48
-20
-6.58
-40
-0.95
-50
-180
-51
-540
-0
-ENDSEC
-0
-EOF
