+++ test/fixtures/bundle/config.yaml
@@ -4,10 +4,6 @@ outlines:
     - what: rectangle
       where: true
       size: 18
-  hotswap:
-    - what: mx_hotswap
-      where: true
-      size: 18
 pcbs:
   pcb:
     outlines.edge.outline: box
