+++ test/outlines/svg_inline___outlines_svgpath_dxf.dxf
@@ -39,7 +39,7 @@ LTYPE
 2
 DASHED
 3
-_ _ _ 
+_ _ _
 73
 2
 40
@@ -57,7 +57,7 @@ LTYPE
 2
 DOTTED
 3
-. . . 
+. . .
 73
 2
 40
