+++ test/outlines/svg_paths_type.yaml
@@ -1,0 +1,12 @@
+points:
+  zones:
+    matrix:
+      columns:
+        pos:
+          key:
+            x: 0
+            y: 0
+outlines:
+  pathserror:
+    - what: svg
+      paths: 123
