+++ test/outlines/svg_paths_type___EXCEPTION.txt
@@ -1,0 +1,1 @@
+must be a string or an array
