+++ test/outlines/svgmirror.yaml
@@ -1,0 +1,15 @@
+points:
+  zones:
+    matrix:
+      columns:
+        pos:
+          key:
+            x: 0
+            y: 0
+  mirror:
+    distance: 20
+outlines:
+  mirror:
+    - what: svg
+      paths: M 0 0 L 10 0 L 10 10 Z
+      where: true
