+++ test/outlines/svgmirror___outlines_mirror_dxf.dxf
@@ -91,23 +91,9 @@ CONTINUOUS
 20
 0
 11
-10
-21
-0
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-10
 20
-0
-11
-10
 21
-5
+0
 0
 LINE
 8
@@ -115,21 +101,7 @@ LINE
 6
 CONTINUOUS
 10
-5
-20
 10
-11
-0
-21
-10
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-0
 20
 10
 11
@@ -145,53 +117,11 @@ CONTINUOUS
 10
 10
 20
-5
-11
-15
-21
-5
-0
-LINE
-8
-0
-6
-CONTINUOUS
 10
-15
+11
 20
-5
-11
-15
 21
-15
 0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-15
-20
-15
-11
-5
-21
-15
-0
-LINE
-8
-0
-6
-CONTINUOUS
-10
-5
-20
-15
-11
-5
-21
-10
 0
 ENDSEC
 0
