+++ test/unit/internals.js
@@ -2,6 +2,7 @@ const m = require('makerjs')
 const kicad5 = require('../../src/templates/kicad5')
 const kicad8 = require('../../src/templates/kicad8')
 const ergogen = require('../../src/ergogen')
+const io = require('../../src/io')
 
 describe('Internals', function() {
 
@@ -27,13 +28,8 @@ describe('Internals', function() {
     })
 })
 
-
-
-
-const io = require('../../src/io')
-
 describe('IO', function() {
-    it('should unpack outlines from zip', async function() {
+    it('should unpack outlines from zip with multiple paths and mirroring', async function() {
         const mockZip = {
             file: function(regex) {
                 if (regex.toString().includes('config')) {
@@ -47,7 +43,7 @@ describe('IO', function() {
                         file: function(regex) {
                             return [{
                                 name: 'outlines/test.svg',
-                                async: () => Promise.resolve('<svg><path d="M 0 0 L 10 0 L 10 10 Z" /></svg>')
+                                async: () => Promise.resolve('<svg><path d="M 0 0 L 10 0 L 10 10 Z" /><path d="M 20 20 L 30 20 L 30 30 Z" /></svg>')
                             }]
                         }
                     }
@@ -61,11 +57,16 @@ describe('IO', function() {
         injections[0][0].should.equal('outline')
         injections[0][1].should.equal('test')
 
-        // Verify the injected function works
+        // Verify the injected function works with multiple paths
         const shapeMaker = injections[0][2]
         const [maker, units] = shapeMaker({}, 'test', {}, {}, {})
-        const [shape, bbox] = maker({ meta: {} })
 
-        shape.paths.should.not.be.empty
+        // Test non-mirrored
+        const [shape, bbox] = maker({ meta: { mirrored: false } })
+        shape.should.exist
+
+        // Test mirrored (covers io.js:81)
+        const [mirroredShape, mirroredBbox] = maker({ meta: { mirrored: true } })
+        mirroredShape.should.exist
     })
 })
